♻️ [Refactoring]: deployment::output モジュール分離 + DeploymentResult→DeploymentOutput リネーム
- src/component/deployment/output.rs に DeploymentOutput / HookConvertOutput / Display impl を抽出
- src/component/deployment/output_test.rs に Display 系 6 件 + HookConvert 系 2 件を移動
- 命名: std::result::Result との混同を避けるため ~Output に統一（std::process::Output 等の慣用に倣う）
- 呼び出し側 (component.rs / install.rs) のパターンマッチ・use 文を更新
- 既存 N=59 件 green 維持
